refactor: separated certification component from account page.
